fastai/backend/requirements.txt
# torch
# torchvision
pillow
numpy
typing
scikit-image
fastapi
python-multipart
uvicorn[standard]
yolov5
opencv-python
